% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_desired.R
\name{calculate_desired_gains}
\alias{calculate_desired_gains}
\title{Calculation of desired gains index (Werner et al.)}
\usage{
calculate_desired_gains(
  A,
  V,
  gains = NULL,
  use.marginal.V = FALSE,
  use.V.approx = FALSE,
  n.Threads = 4L
)
}
\arguments{
\item{A}{A matrix (n_genotype x n_trait) of multi-trait BLUPs.}

\item{V}{Numeric matrix ((n_genotype\emph{n_trait) x (n_genotype}n_trait)) with the
posterior variance var(\eqn{\tilde{g}}).
Required if \code{use.marginal.V} is TRUE.}

\item{gains}{Numeric vector (length = n_trait) of desired gains.}

\item{use.marginal.V}{Logical. If TRUE, uses the \emph{marginal} posterior variance
(mean of per-genotype trait-blocks from \code{V}).}

\item{use.V.approx}{Logical. If TRUE, uses an empirical estimate of V as cov(A).}

\item{n.Threads}{Number of OpenMP threads.}
}
\value{
A list with index values and trait weights.
}
\description{
Calculates the desired gains index based on multi-trait BLUPs and a chosen
(co)variance matrix option.
}
