% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variances_RIL.R
\name{calculate_variances_RIL}
\alias{calculate_variances_RIL}
\title{Segregation variance for recombinant inbred lines from specified crosses}
\usage{
calculate_variances_RIL(
  crosses,
  genetic.map,
  marker.mat,
  effects,
  t,
  intensity,
  covariance = FALSE,
  weights = NULL,
  method = "osthushenrich",
  n.Threads = 4L
)
}
\arguments{
\item{crosses}{A data.frame or matrix (n_crosses x 2) of parental indices
(row indices into \code{M}) specifying the crosses.}

\item{genetic.map}{A data.frame with columns:
\itemize{
\item \code{site}: integer marker index (1..ncol(M))
\item \code{chr}: chromosome identifier (numeric)
\item \code{pos}: position on the chromosome in morgan
}
All markers in \code{M} must appear in \code{genetic.map$site}.}

\item{marker.mat}{Numeric marker matrix (markers in columns, genotypes in rows) with
entries in \code{c(0, 2)} corresponding to \code{c("AA", "BB")}.}

\item{effects}{Numeric matrix of marker effects (markers in rows, traits in columns).
Must have \code{nrow(U) == ncol(M)}.}

\item{t}{Integer. Number of random‐mating generations before RIL creation.}

\item{intensity}{Double. Standardized selection differential.}

\item{covariance}{Logical. If \code{TRUE}, also compute the segregation
covariance matrix between all supplied traits.}

\item{weights}{Numeric vector of length \code{ncol(U)} with fixed trait weights.
If supplied the function calculates index values for each cross. Only works if \code{covariance = TRUE}; otherwise ignored.}

\item{method}{Character. Which method to use; one of
\code{c("osthushenrich", "lehermeier")}.}

\item{n.Threads}{Integer (default 4). Number of OpenMP threads (if enabled at compile time).}
}
\value{
If \code{covariance = TRUE}, a list with element \code{cross_values}
(a data.frame) whose columns are named \code{EGBV1, var1, SPV1, EGBV2, ...} and a list with element \code{covariances}.
If \code{covariance = TRUE} and \code{calculate.index = TRUE}, additionally an element \code{index}
(a data.frame) with \code{IDG_A, VARIDG_A, SPVIDG_A}.
If \code{covariance = FALSE}, a data.frame with the same columns \code{EGBV1, var1, SPV1, EGBV2, ...}.
}
\description{
Calculate expected genomic estimated breeding values (EGBV), segregation
variance, and superior progeny value (SPV) for recombinant inbred lines (RIL)
derived after \code{t} rounds of random mating.
}
