% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossingTools-package.R
\docType{package}
\name{CrossingTools-package}
\alias{CrossingTools}
\alias{CrossingTools-package}
\title{CrossingTools: Mate allocation and cross evaluation helpers}
\description{
Internal package-level documentation stub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/svenernstW/CrossingTools}
  \item Report bugs at \url{https://github.com/svenernstW/CrossingTools/issues}
}

}
\author{
\strong{Maintainer}: Sven Weber \email{sven.e.weber@agrar.uni-giessen.de}

Authors:
\itemize{
  \item Daniel Tolhurst
  \item Dom Waters
  \item Christian Werner
}

}
\keyword{internal}
