% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segvar_inbred.R
\name{get_segvar_inbred}
\alias{get_segvar_inbred}
\title{Segregation variance for inbred lines from specified crosses}
\usage{
get_segvar_inbred(
  crosses,
  genetic.map,
  marker.mat,
  effects,
  t,
  intensity = NULL,
  type = "DH",
  covariance = FALSE,
  weights = NULL,
  method = 1,
  nthreads = 4L
)
}
\arguments{
\item{crosses}{A matrix or data.frame (n_crosses x 2 for two way crosses or n_crosses x 4 for four way crosses) specifying the parents for each
proposed cross. Entries may be either:
\itemize{
\item integer indices referring to rows of \code{marker.mat}, or
\item character identifiers matching \code{rownames(marker.mat)}.
}}

\item{genetic.map}{A data.frame with columns:
\itemize{
\item \code{site}: integer marker index (1..ncol(marker.mat))
\item \code{chr}: chromosome identifier (numeric)
\item \code{pos}: position on the chromosome in morgan
}
All markers in \code{marker.mat} must appear in \code{genetic.map$site}.}

\item{marker.mat}{Numeric marker matrix (markers in columns, genotypes in rows) with
entries in \code{c(0, 2)} corresponding to \code{c("AA", "BB")}.}

\item{effects}{Numeric matrix of marker effects (markers in rows, traits in columns).
Must have \code{nrow(effects) == ncol(marker.mat)}.}

\item{t}{Integer. Number of random‐mating generations before DH creation.}

\item{intensity}{Double. Standardized selection differential, efaults to 1.}

\item{type}{intended offspring typew can be "RIL" for recombinant inbred lines or "DH" for double haploid lines.}

\item{covariance}{Logical. If \code{TRUE}, also compute the segregation
covariance matrix between all supplied traits.}

\item{weights}{Numeric vector of length \code{ncol(effects)} with fixed trait weights.
If supplied the function calculates index values for each cross. Only works if \code{covariance = TRUE}; otherwise ignored.}

\item{method}{Character. Which method to use; one of \code{c(1, 2)} for
Lehermeier or Osthushenrich.}

\item{nthreads}{Integer (default 4). Number of OpenMP threads (if enabled at compile time).}
}
\value{
If \code{covariance = TRUE}, a list with element \code{cross.df}
(a data.frame) whose columns are named \code{EGBEV1, var1, SPV1, EGEBV2, ...} and a list with element \code{covariances}.
If \code{covariance = TRUE} and \code{calculate.index = TRUE}, additionally an element \code{index.df}
(a data.frame) with \code{IDX, VAR.IDX, SPV.IDX}.
If \code{covariance = FALSE}, a data.frame with the same columns \code{EGEBV1, var1, SPV1, EGBEV2, ...}.
}
\description{
Calculate expected genomic estimated breeding values (EGEBV), segregation
variance, and superior progeny value (SPV) for doubled haploid or recombinant inbred lines
#' derived after \code{t} rounds of random mating. works for two way, three way and four way crosses
}
