% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backsolve_marker_effects.R
\name{backsolve_marker_effects}
\alias{backsolve_marker_effects}
\title{Backsolve marker effects from individual effects}
\usage{
backsolve_marker_effects(marker.mat, G.mat, genotype.effects)
}
\arguments{
\item{marker.mat}{Numeric matrix (n x p): marker/genotype matrix (individuals in rows, markers in columns).}

\item{G.mat}{Numeric matrix (n x n): relationship matrix among individuals.}

\item{genotype.effects}{Numeric (n x k) or length-n vector: individual effects (one or more traits).}
}
\value{
A numeric \code{p x k} matrix of marker effects .
}
\description{
Computes marker effects \eqn{\mu = (scalingFactor * M' * G^{-1}) * g}. The scaling factor is automatically determined from the data.
}
