% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_desired.R
\name{calculate_desired_gains}
\alias{calculate_desired_gains}
\title{Calculation of desired gains index (Werner et al.)}
\usage{
calculate_desired_gains(effects, V = NULL, gains = NULL, n.Threads = 4L)
}
\arguments{
\item{effects}{A matrix (n_genotype x n_trait) of multi-trait BLUPs.}

\item{V}{Numeric matrix ((n_genotype\emph{n_trait) x (n_genotype}n_trait)) with the
posterior variance var(\eqn{\tilde{g}}) of the provided effects. If provided,
the \emph{marginal} posterior variance (mean of per-genotype trait-blocks from \code{V})
is used to calculate the index. If not provided an empirical estimate of V as cov(A)
is used.}

\item{gains}{Numeric vector (length = n_trait) of desired gains.}

\item{n.Threads}{Number of OpenMP threads.}
}
\value{
A list with index values and trait weights.
}
\description{
Calculates the desired gains index based on multi-trait BLUPs and a chosen
(co)variance matrix option.
}
\keyword{internal}
