% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cross_plan.R
\name{make_cross_plan}
\alias{make_cross_plan}
\title{Creation of a plan of all potential crosses for CrossingTools}
\usage{
make_cross_plan(
  candidates = NULL,
  male_candidates = NULL,
  female_candidates = NULL
)
}
\arguments{
\item{candidates}{A vector of integers that identify each genotype. Should be consistent with genotypic data.
Use this function if there are no sexes.}

\item{male_candidates}{A vector of integers that identify each male genotype. Ignored if candidates is supplied.
Should be consistent with genotypic data.}

\item{female_candidates}{A vector of integers that identify each female genotype. Ignored if candidates is supplied.
Should be consistent with genotypic data.}
}
\value{
A two-column data.frame with the the combinations of all supplied individuals.
If male and female candidates are supplied, the first column corresponds to males, and the second to females
}
\description{
Creation of a plan of all potential crosses for CrossingTools
}
