% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_cross_selection.R
\name{optimal_cross_selection}
\alias{optimal_cross_selection}
\title{Optimal cross selection via a genetic algorithm}
\usage{
optimal_cross_selection(
  ncrosses,
  target.angle,
  g,
  G,
  propability.mutate = 0.01,
  nmutate = 2,
  nselect = 500,
  npop = 10000,
  max.generation = 100,
  max.iteration = 100,
  angle.penalty,
  nthreads = 4L
)
}
\arguments{
\item{ncrosses}{Integer. Number of crosses to select.}

\item{target.angle}{Numeric (radians). Target angle between gain and diversity.}

\item{g}{Numeric vector of per-potential-cross scores (length must be \eqn{nInd*(nInd-1)/2}).
Each element corresponds to one half-diallel cross.}

\item{G}{Matrix square relationship matrix among individuals (\eqn{nInd \times nInd}).}

\item{propability.mutate}{Numeric in \code{[0,1]}. Per-solution mutation probability in GA.}

\item{nmutate}{Integer \(\ge 0\). Number of potential mutations when a mutation occurs.}

\item{nselect}{Integer \(\ge 2\). Number of solutions retained per generation in GA.}

\item{npop}{Integer \(\ge 2\). Solution population size (solutions per generation).}

\item{max.generation}{Integer \(\ge 1\). Maximum GA iterations.}

\item{max.iteration}{Integer \(\ge 1\). Early-stop if no improvement for this many solution generations.}

\item{angle.penalty}{Numeric \(\ge 0\). Penalty weight for deviating from \code{target.angle}.}

\item{nthreads}{Integer \(\ge 1\). number of threads to use.}
}
\value{
A list:
\describe{
\item{crossPlan}{\code{ncrosses x 2} matrix of selected crosses.}
\item{uMax, uMin}{Best/achieved mean cross value}
\item{simMax, simMin}{Best/achieved similarity/relationship.}
\item{uBest, simBest}{Mean cross value and similarity of the final best solution.}
\item{angleBest, lenBest}{Angle and vector length of the final best solution.}
}
}
\description{
Runs a GA to select a set of \code{ncrosses} crosses that balance
expected gain (\code{g}) and genetic similarity/diversity (via \code{G})
around a target trade-off angle.
}
\examples{
\dontrun{
nInd <- 50
G <- diag(nInd) # toy
potCross <- nInd*(nInd-1)/2
g <- rnorm(potCross)
res <- optimal_cross_selection(
  ncrosses = 20, target.angle = pi/4, g = g, G = G,
  propability.mutate = 0.01, nmutate = 2, nselect = 200,
  npop = 2000, max.generation = 200, max.iteration = 50,
  angle.penalty = 0.5, nthreads = 4L
)
}
}
